-interface for ImagesStoreService created (AmazonClient is implementation) -runtime exception MemberRequestException now handles bad requests -check of fileUrl not null during delete in Amazon client -security tests fixed

+++ src/main/java/com/ncube/member/services/AmazonClient.java
@@ -3,12 +3,12 @@ package com.ncube.member.services;
 import com.amazonaws.auth.AWSCredentials;
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
-import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
 import com.amazonaws.services.s3.model.CannedAccessControlList;
 import com.amazonaws.services.s3.model.DeleteObjectRequest;
 import com.amazonaws.services.s3.model.PutObjectRequest;
+import com.ncube.member.MemberRequestException;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -20,7 +20,7 @@ import java.io.IOException;
 import java.util.Objects;
 
 @Service
-public class AmazonClient {
+public class AmazonClient implements ImagesStoreService<MultipartFile> {
 
     private AmazonS3Client s3client;
 
@@ -41,14 +41,15 @@ public class AmazonClient {
                 .withCredentials(new AWSStaticCredentialsProvider(credentials)).build();
     }
 
+    @Override
     public String uploadFile(MultipartFile multipartFile, String uploadFileName) {
-        String fileUrl = "";
+        String fileUrl;
         try {
             File file = convertMultiPartToFile(multipartFile);
             fileUrl = uploadFileTos3bucket(uploadFileName, file);
             file.delete();
         } catch (IOException e) {
-            //TODO
+            throw new MemberRequestException(String.format("Error during saving image : %s", e.getMessage()));
         }
         return fileUrl;
     }
@@ -67,8 +68,11 @@ public class AmazonClient {
         return s3client.getResourceUrl(bucketName, fileName);
     }
 
-    public void deleteFileFromS3Bucket(String fileUrl) {
-        String fileName = fileUrl.substring(fileUrl.lastIndexOf("/") + 1);
-        s3client.deleteObject(new DeleteObjectRequest(bucketName, fileName));
+    @Override
+    public void deleteFile(String fileUrl) {
+        if (fileUrl != null) {
+            String fileName = fileUrl.substring(fileUrl.lastIndexOf("/") + 1);
+            s3client.deleteObject(new DeleteObjectRequest(bucketName, fileName));
+        }
     }
 }
